refactor: adjust code about frontend
