+++ compose.yml
@@ -30,7 +30,7 @@ services:
       MYSQL_TEST_DATABASE: 'gotea_test'
       TZ: 'Asia/Tokyo'
     ports:
-      - '33306:3306'
+      - '3306:3306'
     command: mysqld --character-set-server=utf8mb4 --collation-server=utf8mb4_general_ci
 
 volumes:
