+++ resources/assets/scripts/libs/validator.ts
