+++ resources/assets/scripts/vuex.d.ts
@@ -1,7 +1,8 @@
 import { Store } from 'vuex';
+import { State } from '@/types';
 
 declare module '@vue/runtime-core' {
   interface ComponentCustomProperties {
-    $store: Store<T>;
+    $store: Store<State>;
   }
 }
